+++ src/config/toml.rs
@@ -36,6 +36,7 @@ pub struct Toml {
     pub header: Option<bool>,
     pub utc: Option<bool>,
     pub color: Option<String>,
+    pub output: Option<String>,
     pub log: Option<TomlLog>,
     pub predict: Option<TomlPred>,
     pub stats: Option<TomlStats>,